- property sensor was converting floating point values to text then back to float - for floating point properties. - IntValue's GetNumber() was convert int -> float -> double. - BL_Shader was using STR_String rather then char*, where most callers had a char*, use a char* to avoid STR_String conversion-and-alloc on shader access.

+++ source/gameengine/Expressions/IntValue.cpp
@@ -286,7 +286,7 @@ cInt CIntValue::GetInt()
 
 double CIntValue::GetNumber()
 {
-	return (float) m_int;
+	return (double) m_int;
 }
